No special case for the `set` transition.
Microstate transitions are like a transaction. You start with a
microstate, and what is returned is the next wersion of that
microstate.

However, this transaction was being spread over two methods: the
actual transition invocation and the `set` method. In other words, the
transition method was just called, and then the "work" was happening
in `set`.

As we move the codebase to have more interesting outcomes inside a
single transition like being able to transition to different parts of
the store, we'll need to be able to reason about what exactly a state
transition is in a single place.

This change places all of the transition logic in a single location:
the transition method, and `set` becomes just another transition that
isn't at all special except for the fact that a default
implemenation (the identity function) is provided for it without you
having to specify it in the class definition.

Having no special transitions will give us the power to make changes
to transition semantics without having to worry about corner cases.