Added object iterator to allow using ObjectType with queries (#276)
Introduces `Symbol.iterator` for ObjectType. This makes it possible to use object microstates with `for (let obj of microstates)` syntax, which also provides support for Queries to map, filter and reduce objects.

The iterator yields entry objects which allow the key/value to be restructured as objects `for (let { key, value } of microstates)` or arrays `for (let [value key] of microstates)`.