Gather all transitions from the prototype chain (#290)
When a transition is inherited from the parent, the transition
definition resides on the parent's prototype. This resulted in
closures not being created around parent's transitions because
methodOf function that was used to gather transitions was only looking
on the class's prototype.

This PR introduces getPrototypeDescriptors function that gathers
prototypes by recursively walking the prototype chain. methodOf
consumes this function to ensure that all transitions are included
when creating closures for transitions.