Don't allow microstates to become part of Array values.
With Union types (which are in the works) a very common use-case is to
pass in a microstate union in a transition on a parent microstate:

```js
uploader.uploads.push(Upload.New.create({ file }))
````

What we really want is the _value_ of the upload, not the actual
microstate upload itself. That means that whenever we want to include
a value into a container object like an array, then we need to unwrap
the microstate before putting it into the value. This ensures that the
microstate itself can always be re-hydrated from its simple POJO
values.

This unwraps the value before putting it into an array value for both
push and pop. Any future array method that creates a record, like insert for
example, should also be careful only to include simple values.