Use abstract operation to access microstate children by key.
An identity is a reference to a microstate that can change over
time, but the actual Id object doesn't have a direct reference to the
microstate it represents. Instead it holds the _path_ of the
microstate and looks it up each time from the latest value that is
contained in the identity. This process of looking up the
corresponding microstate means navigating down each element of the
path.

But this presents a problem. Due to the laziness of Arrays (and soon
Objects) We can't use standard keyed access all the time because array
microstates can only access their children via enumeration. In other
words, normal JS property access doesn't work.

```js
let array = create([Number], [1,2,3]);
array[2] //=> undefined

let [one, two] = array;
one //=> Microstate<Number>;
```

We got around this before by putting in a special hack for arrays to
check "hey, is this an array microstate? If so, let's give a separate
treatment." Not only was this cumbersome, but we'll need to add yet
another hack for `ObjectType`, and furthermore, it is incompatible
with the changes that we need to put in place to make identities
support union types.

This change adds a typeclass `Tree` with a single method `childAt`
which abstracts away the JS operation for property access. That way,
if we have the path of an entity, we can resolve it even if it's
comprised of many different types of objects that may or may not
support direct keyed access.

As a result, the `At` lens now uses `childAt`, so that microstates can
also be traversed, not just POJOs. However, only for the `view`
operation. A corresponding polymorphic `setChildAt()` would be
required to make a change.