Mke transitions in a store return latest node at transition path.
Right now, a microstate created “in the wild” with the `create()`
function always returns an instance of that microstate for all of its
transitions and all of its child transitions.

```js
create(class Counter { count = Number; })
  .set({ count: 5 })
  .count.increment()
  .count.increment(-3)
```

This makes sense within the context of a transition, where you’re
always dealing with the root node of the thing being transitioned.
Microstates inside a store follow this pattern currently, but I’m not
so sure it’s a good idea.  The reason is that most of the time, when
you’re invoking a transition from a microstate referenced in a store,
you:

1. are not invoking a chain, you’re invoking the outermost transition
2. you’re working with a “slice” of the tree, not the entire tree

So this API isn’t as useful, and it could even be bad. The reason this
occured to me is in the modelling of side-effects.

For example, imagine you have a (contrived) microstate tree that looks
like this:

```class App {
  router = class Router {
    avatars = class AvatarPage {
      uploader = Uploader {
        uploads = [Upload];
        editor = Editor;
    };
  };
}
```

and I want to have some effect-ful logic that operates just on the
file uploads. For an example, see:

https://github.com/cowboyd/microstates-file-upload/blob/master/src/uploader-controller.js#L63-L77)

In microstates today, every transition returns the root of the
tree. In other words, if this uploader were embedded into the `App`
object like above, then the result of `upload.start()` would be an
`App` object:

```js
upload.start() //=> App {}
```

Not only is this not really useful, it’s potentially dangerous in that
we’re effectively “leaking” the enclosing context into a
sub-context. What if we wanted to take our App, and put it into a
multi-tenant platform like:

```js
class Platform {
  apps = [App];
}
```

If you had an upload that was somehow depending on the fact that
`upload.start()` returned an `App` microstate, it would break when
suddenly it now returns a `Platform` microstate.  and we _always_ want
to be able to embed microstates into other microstates without
negative consequences.

Pursuant to the discussion on slack, this implements option (1) where
transitions always return the "freshest" version of the microstate at
a given path.

```js

upload //=> Upload.New {}
upload.start() //=> Upload.Started {}
```