Remove useless `promap`
Back when the microstate tree was eagerly evaluated, we had to
whenever we mounted one tree onto another we had to map over the
entire tree in memory and update the metadata associated with each
node in order to adjust its path and so forth.

But now that the tree is evaluated completely lazily, this type of
mapping is no longer necessary. Instead it is enough to derive the
mounted substate from the current state which is its parent, and all
subsequent substates will do the same.

Note that this is the first step towards removing the `Profunctor`
class altogether.