Remove Profunctor class entirely.
I had planned to remove this on a different branch, but coveralls
is (correctly) complaining about all this dead code.

This goes ahead and removes it on this branch so that the coverage
numbers aren't thrown off.

In the mean time, this removes some obsolete static symbols for
metadata that is no longer stored.