Don't try to reify a value if the iteration is finished.
For each `step` of an iteration, the iterator has two properties:
`step.done` and `step.value`. In the case of arrays and objects, if
the `step.done` method returns false, the expectation is that you've moved
past the end of list collection and that you won't even bother with
calling the `step.value` because why? And this is mostly true of
libraries that use iterator.

However, in the case of generator functions, each `step` represents
the invocation of `yield`, and the _last_ step, in which `step.done`
is `true` represents the return value of the generator function. In
other words, `step.value` _can_ have a meaningful value. So for
libraries that generalize over all iterators, not just collection
iterators (like regenerator-runtime), we need to have `value` return
something meaningful, not just a garbage entry with an invalid value.

Luckily, this is straightforward. For both `ObjectType`, and
`ArrayType` we guard the iteration step value by checking first to see
if the iteration is done, and return `undefined` if we've moved past
the end of the collection.