Alaways derive store location value from its microstate value.
When a `Location` was created in the pathmap, it was using the value
at the path in the raw POJO as the value of the Reference. However,
because of microstates allowing default values, the actual value of a
microstate _can_ be different than the value passed into the
constructor.

In the case where you have _undefined_  as the value of the
store, the raw value in the pojo is also undefined, but the value at
the microstate at that path is the default.

Therefore, we need to actually use the value of the _microstate_ at
that path as the value of the location, so that references will use
_that_ value instead.

With this change, the `currentValue` of the `Location` looks up the
path in the microstate instead of the POJO and returns that value
instead.

Fixes https://github.com/microstates/microstates.js/issues/329