Don't store keys on the object type.
Objects are a bit weird in that they actually keep all of the keys as
properties of the object microstate. This has a couple of problems:

Performance
====

Because we need to iterate over all of the keys of the object value
and define a property for them, it means a significant performance
hit. Even though the values themselves are built lazily, the property
definitions need to be created on every single construction

Naming conflicts
====

Let's say you have an object that has keys with same names as
transitions, like `{ put: 'the plate on the table' }`. This means that
the _property_ "put" will override the _transition_ "put".

This seems untenable long term since it applies to every transition,
including `initialize`, `delete`, `filter`..... all of them. The
microstate would most likely break completely if the value had a
property called "initialize".

This removes the direct properties from the Object microstate. They
are still available as the `entries` property, but now they are
completely contained within their own namespace and can't interfere
with the Object namespace.

In practice, I expect that the `entries` property will be fairly
rarely used, and in actual application, the way developers will access
the substates in an object will be through enumeration.