Abstract relationships.
How to contextualize a child microstate has been a constant
challenge. In other words, sometimes you want a child to take on a
constant value. Sometimes you want it to take on a default value, and
sometime you might want it to be completly polymorphic on what value
it takes on depending on the parent.

Unfortunately, this is not possible with our current way of expressing
relationships with is either with a DSL:

```js
class Counter {
  count = Number;
}
```

or with an explicitly created microstate:

```js
class Counter {
  count = create(Number, 1);
}
```

The problem is that there is no way to expand the concept of what gets
created if you want to do something fancier, like copy values down the
graph without resorting to intializer hacks.

This introduces a very thin concept of a _relationship_ between a parent
microstate and a child microstate. The responsibility of the traversal
is to take the value at the location of the target microstate and
return a specifier for the child microstate.

To do this, we introduce the concept of a "Cell" which is a Type,
value pair `{ Type, value }`. This is the minimal about of information
required to specify a microstate. This is super light-weight because
it does not require the overhead of actually `create`-ing a
microstate.

It is the responsibility of a relationship to resolve a value into a
cell.

> Note: in future microstates, we will use `{ Type, path }` as a cell
  specifier.

The relationship class is organized monadically so that it can be
mapped, flatMapped, etc... and any relationship can be extended using
`map` and `flatMap`

For example, to implement a hypothetical `child` helper which creates
a specifies a substate of a given type and default value, it could
look like this.

```js
function child(Type, defaultValue) {
  return relationship()
    .map(({ value }) => value != null ? {Type, value} : {Type, value: defaultValue })
}
```