🐛Make sure to build before using as an NPM dependency
Microstates uses rollup to transpile its source and make it available
in a number of different module formats. It won't work unless this
step happens before a publish.

We can use the `prepack` [npm script][1] to make sure that if someone
is generating a tarball, publishing, or installing as a `git`
dependency that the build gets run.

[1]: https://docs.npmjs.com/misc/scripts