Remove marketing message
We had hoped that explicitly calling out our consulting services migth
be a driver for business, so we added a blatant marketing message to
the top of the readme.

It was definitely worth the shot, especially give the situation in
which we found ourselves (needing to leave no stone
unturned). However, that hasn't really panned out, and we're a little
bit more low-key and happy to let the "created by frontside" badge
speak for itself.