✍🏼Document the `from()` function to account for collection laziness
A while back, the change was made to not allow direct keyed access to
[Object][1] and [Array][2]. While we updated most of the documentation
to include those changes, the documentation of the `from()` method was
not. As a result, this was misleading.

This change refactors the examples to use either the `entries` map on
object or the destructuring which uses the iterability of both Array
ond Object.

[1]: https://github.com/microstates/microstates.js/pull/338
[2]: https://github.com/microstates/microstates.js/pull/227